// frontend/src/app/models/employee.ts
export interface Employee {
  _id?: string;
  employeeId: string;
  firstName: string;
  lastName: string;
  dob: string;
  gender: string;
  email: string;
  phone: string;
  department: string;
  jobTitle: string;
  employmentType: string;
  hireDate: string;
  employeeStatus: string;
  supervisor: string;
  salary: number;
  payFrequency: string;
  bankAccount: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  emergencyContactName: string;
  emergencyContactNumber: string;
  notes: string;
}